TWPA power - reduced - URGENT
On the qw11q the lab installed a different TWPA which operates at a much lower power. As a safety feature I suggest to reduce the power to avoid possible damage of the twpa. Note that at the moment the TWPA is not calibrated and the purpose of this is to avoid any potential harm to the twpa.